+++ Events/TouchdownEvent.cs
@@ -21,7 +21,7 @@ namespace EddiEvents
         public string bodyname { get; private set; }
 
         [PublicAPI( "The numeric ID of the body where the ship has touched down" )]
-        public long? bodyId { get; private set; }
+        public int? bodyId { get; private set; }
 
         [PublicAPI("The longitude from where the ship has touched down")]
         public decimal? longitude { get; private set; }
@@ -52,7 +52,7 @@ namespace EddiEvents
 
         public bool? onplanet { get; private set; } // Always true, since `Touchdown` is currently only ever triggered when touching down on a body
 
-        public TouchdownEvent(DateTime timestamp, decimal? longitude, decimal? latitude, string system, ulong systemAddress, string body, long? bodyId, bool? onStation, bool? onPlanet, bool? taxi, bool? multicrew, bool playercontrolled, SignalSource nearestDestination) : base(timestamp, NAME)
+        public TouchdownEvent(DateTime timestamp, decimal? longitude, decimal? latitude, string system, ulong systemAddress, string body, int? bodyId, bool? onStation, bool? onPlanet, bool? taxi, bool? multicrew, bool playercontrolled, SignalSource nearestDestination) : base(timestamp, NAME)
         {
             this.longitude = longitude;
             this.latitude = latitude;
